Button Delete into settings(Edit route)
